troubles with the my-tv-shows-find-missing task
i have experienced a silent crash of flexget twice in the last 2 days... i think the backlog plugin was getting too full with the find-missing task running
so often. i have changed from the hourly schedule to an interval of 3 days to see if that will work; as per Jeff's original config.