maybe that didn't work
i thoughtt hat the difference might have just been that <insert guy's name
here> was using both username and account tags in his trakt plugins... not
so much... i have commented out the things that didnt work... now to crack
the case on the RSS feed. this is gonna be totallly new to me.